Backend/feat/92+71 소켓통신 좌표정보 db저장+snapshot delete (#93)
* feat : 스냅샷 생성 갯수 제한 15개

* chore : origin 허용 수정

* feat : 노트 정보 소켓 통신 시 DB에 영구 반영되도록 구현
spaceId 하드 코딩 되있는거 변수 전달로 변경

* chore : 악기 String -> Enum 변환 추가

* chore : 악기 좌표 Integer로 변경

* feat : 스냅샷 삭제 후 workspace의 snapshot 개수 -1

+++ backend/letsnote/src/main/java/com/geeks/letsnote/domain/studio/workSpace/dto/ResponseNotes.java
@@ -14,8 +14,8 @@ public record ResponseNotes() {
 
     @Builder
     public record Note (
-            Long noteX,
-            Long noteY
+            Integer noteX,
+            Integer noteY
     ) {}
